Extract timestamp functions/operators whose semantics is legacy
