+++ presto-main/src/main/java/com/facebook/presto/type/DateOperators.java
@@ -101,18 +101,6 @@ public final class DateOperators
     }
 
     @ScalarOperator(CAST)
-    @SqlType(StandardTypes.TIMESTAMP)
-    public static long castToTimestamp(ConnectorSession session, @SqlType(StandardTypes.DATE) long value)
-    {
-        long utcMillis = TimeUnit.DAYS.toMillis(value);
-
-        // date is encoded as milliseconds at midnight in UTC
-        // convert to midnight if the session timezone
-        ISOChronology chronology = getChronology(session.getTimeZoneKey());
-        return utcMillis - chronology.getZone().getOffset(utcMillis);
-    }
-
-    @ScalarOperator(CAST)
     @SqlType(StandardTypes.TIMESTAMP_WITH_TIME_ZONE)
     public static long castToTimestampWithTimeZone(ConnectorSession session, @SqlType(StandardTypes.DATE) long value)
     {
