+++ presto-main/src/main/java/com/facebook/presto/type/TimestampOperators.java
@@ -13,23 +13,13 @@
  */
 package com.facebook.presto.type;
 
-import com.facebook.presto.spi.ConnectorSession;
-import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.function.IsNull;
-import com.facebook.presto.spi.function.LiteralParameters;
-import com.facebook.presto.spi.function.ScalarFunction;
 import com.facebook.presto.spi.function.ScalarOperator;
 import com.facebook.presto.spi.function.SqlType;
 import com.facebook.presto.spi.type.AbstractLongType;
 import com.facebook.presto.spi.type.StandardTypes;
-import io.airlift.slice.Slice;
-import org.joda.time.chrono.ISOChronology;
 
-import java.util.concurrent.TimeUnit;
-
-import static com.facebook.presto.spi.StandardErrorCode.INVALID_CAST_ARGUMENT;
 import static com.facebook.presto.spi.function.OperatorType.BETWEEN;
-import static com.facebook.presto.spi.function.OperatorType.CAST;
 import static com.facebook.presto.spi.function.OperatorType.EQUAL;
 import static com.facebook.presto.spi.function.OperatorType.GREATER_THAN;
 import static com.facebook.presto.spi.function.OperatorType.GREATER_THAN_OR_EQUAL;
@@ -39,13 +29,6 @@ import static com.facebook.presto.spi.function.OperatorType.LESS_THAN;
 import static com.facebook.presto.spi.function.OperatorType.LESS_THAN_OR_EQUAL;
 import static com.facebook.presto.spi.function.OperatorType.NOT_EQUAL;
 import static com.facebook.presto.spi.function.OperatorType.SUBTRACT;
-import static com.facebook.presto.spi.type.DateTimeEncoding.packDateTimeWithZone;
-import static com.facebook.presto.type.DateTimeOperators.modulo24Hour;
-import static com.facebook.presto.util.DateTimeUtils.parseTimestampWithoutTimeZone;
-import static com.facebook.presto.util.DateTimeUtils.printTimestampWithoutTimeZone;
-import static com.facebook.presto.util.DateTimeZoneIndex.getChronology;
-import static io.airlift.slice.SliceUtf8.trim;
-import static io.airlift.slice.Slices.utf8Slice;
 
 public final class TimestampOperators
 {
@@ -107,63 +90,6 @@ public final class TimestampOperators
     public static boolean between(@SqlType(StandardTypes.TIMESTAMP) long value, @SqlType(StandardTypes.TIMESTAMP) long min, @SqlType(StandardTypes.TIMESTAMP) long max)
     {
         return min <= value && value <= max;
-    }
-
-    @ScalarFunction("date")
-    @ScalarOperator(CAST)
-    @SqlType(StandardTypes.DATE)
-    public static long castToDate(ConnectorSession session, @SqlType(StandardTypes.TIMESTAMP) long value)
-    {
-        // round down the current timestamp to days
-        ISOChronology chronology = getChronology(session.getTimeZoneKey());
-        long date = chronology.dayOfYear().roundFloor(value);
-        // date is currently midnight in timezone of the session
-        // convert to UTC
-        long millis = date + chronology.getZone().getOffset(date);
-        return TimeUnit.MILLISECONDS.toDays(millis);
-    }
-
-    @ScalarOperator(CAST)
-    @SqlType(StandardTypes.TIME)
-    public static long castToTime(ConnectorSession session, @SqlType(StandardTypes.TIMESTAMP) long value)
-    {
-        return modulo24Hour(getChronology(session.getTimeZoneKey()), value);
-    }
-
-    @ScalarOperator(CAST)
-    @SqlType(StandardTypes.TIME_WITH_TIME_ZONE)
-    public static long castToTimeWithTimeZone(ConnectorSession session, @SqlType(StandardTypes.TIMESTAMP) long value)
-    {
-        int timeMillis = modulo24Hour(getChronology(session.getTimeZoneKey()), value);
-        return packDateTimeWithZone(timeMillis, session.getTimeZoneKey());
-    }
-
-    @ScalarOperator(CAST)
-    @SqlType(StandardTypes.TIMESTAMP_WITH_TIME_ZONE)
-    public static long castToTimestampWithTimeZone(ConnectorSession session, @SqlType(StandardTypes.TIMESTAMP) long value)
-    {
-        return packDateTimeWithZone(value, session.getTimeZoneKey());
-    }
-
-    @ScalarOperator(CAST)
-    @LiteralParameters("x")
-    @SqlType("varchar(x)")
-    public static Slice castToSlice(ConnectorSession session, @SqlType(StandardTypes.TIMESTAMP) long value)
-    {
-        return utf8Slice(printTimestampWithoutTimeZone(session.getTimeZoneKey(), value));
-    }
-
-    @ScalarOperator(CAST)
-    @LiteralParameters("x")
-    @SqlType(StandardTypes.TIMESTAMP)
-    public static long castFromSlice(ConnectorSession session, @SqlType("varchar(x)") Slice value)
-    {
-        try {
-            return parseTimestampWithoutTimeZone(session.getTimeZoneKey(), trim(value).toStringUtf8());
-        }
-        catch (IllegalArgumentException e) {
-            throw new PrestoException(INVALID_CAST_ARGUMENT, "Value cannot be cast to timestamp: " + value.toStringUtf8(), e);
-        }
     }
 
     @ScalarOperator(HASH_CODE)
